requirements.txt
asgiref==3.11.1
beautifulsoup4==4.14.3
certifi==2026.1.4
charset-normalizer==3.4.4
contourpy==1.3.3
cycler==0.12.1
Django==6.0.2
django-cors-headers==4.9.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
filelock==3.20.0
fonttools==4.61.1
fsspec==2025.12.0
gdown==5.2.1
idna==3.11
imutils==0.5.4
Jinja2==3.1.6
kiwisolver==1.4.9
MarkupSafe==2.1.5
matplotlib==3.10.8
mpmath==1.3.0
networkx==3.6.1
numpy==2.3.5
opencv-python==4.13.0.90
packaging==26.0
pillow==12.0.0
polars==1.37.1
polars-runtime-32==1.37.1
psutil==7.2.2
PyJWT==2.11.0
pyparsing==3.3.2
PySocks==1.7.1
python-dateutil==2.9.0.post0
PyYAML==6.0.3
RapidFuzz==3.14.3
requests==2.32.5
scipy==1.17.0
setuptools==70.2.0
six==1.17.0
soupsieve==2.8.3
sqlparse==0.5.5
sympy==1.14.0
torch==2.10.0+cpu
torchvision==0.25.0+cpu
tqdm==4.67.3
typing_extensions==4.15.0
ultralytics==8.4.11
ultralytics-thop==2.0.18
urllib3==2.6.3
